Extract all the block bodies (transactions and uncles)
The block bodies are not stored in a completely straightforward way in Erigon.
The transactions are in a separate table, so they can be looked up
individually.  Uncles are included in this extraction.

This may not be quite the right format for our purposes.  It is mainly to be
able to get the size of the data, and to perform compression experiments on it
to see which methods work well.

This version extracts duplicate block numbers if they are present, rather than
filtering to the canonical chain blocks.  Duplicate blocks appear in clusters
surprisingly far in the past.  Those are not useful, and are incorrect if the
reader is assuming these blocks form a single chain, so the non-canonical
blocks should be filtered out.