Accept incarnation may decrease from account to storage in `PlainState`
Improve the error messages when failing sanity checks about the incarnation
value seen in storage slots from the Erigon DB.  It now shows the full block,
address, storage context.

Downgrade these errors to warnings, because they occur a few times in the
Goerli data.  It's not clear if this is correct in the source data, but accept
it for now.