Partition the state-history to independent pages for O(log N) search
This subtle change makes the file O(log N) searchable, quite fast in practice.

- Any encoded entry that would be split over a page boundary is replaced by
  page-padding zeros at the end of the page, and encoded at the start of the next
  page.

- The encoding "previous" state is reset at the start of the next page.

This allows individual entries to be found by searching a single page, when the
correct page is found.  All together, making an O(log N) time query.  (If
interior pages are used and cached well, it effectively becomes O(1) time.)

The page gap is not strictly necessary, because it's possible to read two pages
when querying.  The gap adds a little space overhead, so why have it?

Query performance of _random-reads_ from data too large to fit in RAM depends
on pure random IOPS and bandwidth.  Reading one aligned page is slightly faster
than two, in the best case when they are contiguous on the storage medium, and
about twice as fast when they are not contiguous.

In our application, those kinds of queries are a bottleneck.

It turns out the page gap is smaller than overhead than the encoding state
reset causes indirectly, by making the first entry grow.  We need those resets
spaced not too far apart to set a boundary from where we can start scanning for
reasonably fast queries.  So adding the page gap doesn't add much more; we may
as well have it for faster I/O.  (That said, when a page is reached from a
parent page pointer, context from the parent is available to reduce the size of
the child's first entry.)

The page gap overhead is also not entirely optional.  If the gap was not there
to make the next page's first entry page-aligned, pointers to that entry from
parent pages would need more bits to identify the start offset, or some way of
marking the first entry.

All together, it's worth having the page gap for small entries, and worth
splitting large entries (no entries are that large in the current format).